QuanLyThuVien/QuanLyThuVien/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.9
inflection==0.5.1
ipaddr==2.2.0
jwcrypto==1.5.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
utils==1.0.2
